+++ documentation/src/test/java/net/jqwik/docs/FrequencyOfExamples.java
@@ -2,6 +2,7 @@ package net.jqwik.docs;
 
 import net.jqwik.api.*;
 import net.jqwik.api.arbitraries.*;
+import net.jqwik.api.statistics.Statistics;
 
 class FrequencyOfExamples {