+++ documentation/src/test/java/net/jqwik/docs/InjectDuplicateExamples.java
@@ -4,6 +4,7 @@ import java.util.*;
 
 import net.jqwik.api.*;
 import net.jqwik.api.Tuple.*;
+import net.jqwik.api.statistics.Statistics;
 
 class InjectDuplicateExamples {