+++ documentation/src/test/java/net/jqwik/docs/statistics/StatisticsExamples.java
@@ -1,10 +1,11 @@
-package net.jqwik.docs;
+package net.jqwik.docs.statistics;
 
 import java.math.*;
 import java.util.*;
 
 import net.jqwik.api.*;
 import net.jqwik.api.constraints.*;
+import net.jqwik.api.statistics.Statistics;
 
 class StatisticsExamples {
 
@@ -34,8 +35,9 @@ class StatisticsExamples {
 	}
 
 	@Property
-	void twoParameterStats(@ForAll @Size(min = 1, max = 10) List<Integer> aList, //
-			@ForAll @IntRange(min = 0, max = 10) int index //
+	void twoParameterStats(
+		@ForAll @Size(min = 1, max = 10) List<Integer> aList, //
+		@ForAll @IntRange(min = 0, max = 10) int index //
 	) {
 		Statistics.collect(aList.size() > index ? "index within size" : null);
 	}
@@ -43,18 +45,18 @@ class StatisticsExamples {
 	@Property(tries = 100000)
 	void statsWithRoundedNumbers(@ForAll @IntRange(max = 100000) int anInt) {
 		String classifier = anInt < 10 ? "smallest" //
-				: anInt < 1000 ? "small" //
-						: anInt < 80000 ? "normal" //
-								: "big";
+								: anInt < 1000 ? "small" //
+									  : anInt < 80000 ? "normal" //
+											: "big";
 		Statistics.collect(classifier);
 	}
 
 	@Property(tries = 100000)
 	void statsWithDecimalNumbers(@ForAll int anInt) {
 		String classifier = anInt == 3 ? "three" //
-				: anInt < 1000 ? "small" //
-						: anInt < 80000 ? "normal" //
-								: "big";
+								: anInt < 1000 ? "small" //
+									  : anInt < 80000 ? "normal" //
+											: "big";
 		Statistics.collect(classifier);
 	}