+++ engine/src/main/java/net/jqwik/engine/facades/StatisticsFacadeImpl.java
@@ -2,9 +2,9 @@ package net.jqwik.engine.facades;
 
 import java.util.*;
 
-import net.jqwik.api.*;
-import net.jqwik.api.Statistics.*;
 import net.jqwik.api.lifecycle.*;
+import net.jqwik.api.statistics.*;
+import net.jqwik.api.statistics.Statistics.*;
 import net.jqwik.engine.hooks.statistics.*;
 
 /**